Move `AccountCache` to `ServiceEndpoint`
