+++ android/src/main/kotlin/net/mullvad/mullvadvpn/service/MullvadVpnService.kt
@@ -254,6 +254,7 @@ class MullvadVpnService : TalpidVpnService() {
                 endpoint.messenger,
                 daemon,
                 daemonInstance.intermittentDaemon,
+                endpoint.accountCache,
                 connectionProxy,
                 customDns,
                 endpoint.settingsListener,
