+++ android/src/main/kotlin/net/mullvad/mullvadvpn/service/ServiceInstance.kt
@@ -1,6 +1,7 @@
 package net.mullvad.mullvadvpn.service
 
 import android.os.Messenger
+import net.mullvad.mullvadvpn.service.endpoint.AccountCache
 import net.mullvad.mullvadvpn.service.endpoint.SettingsListener
 import net.mullvad.mullvadvpn.util.Intermittent
 
@@ -8,15 +9,13 @@ class ServiceInstance(
     val messenger: Messenger,
     val daemon: MullvadDaemon,
     val intermittentDaemon: Intermittent<MullvadDaemon>,
+    val accountCache: AccountCache,
     val connectionProxy: ConnectionProxy,
     val customDns: CustomDns,
     val settingsListener: SettingsListener,
     val splitTunneling: SplitTunneling
 ) {
-    val accountCache = AccountCache(settingsListener, intermittentDaemon)
-
     fun onDestroy() {
-        accountCache.onDestroy()
         connectionProxy.onDestroy()
         customDns.onDestroy()
     }
